open and close lowerdir fd around main loop
Hold open a file descriptor on the lower directory in which we
store our non-projected files.  In subsequent commits we'll change
over to using this fd with the openat() family of functions, which
will allow us to eventually remove the need for a chdir() and fixed
working directory.

Since we expect our library's caller to want to chdir() for their
own purposes, this change will support that functionality.

It may also allow (with some caveats) for absolute path to our
lower directory to change without triggering problems in our library.