convert getattr op to use of lowerdir fd
Also add a helper lowerpath() function which we'll use in most
other file operations; this removes the leading / character we
expect from FUSE on our filesystem's paths, to create paths
relative to the lower directory file descriptor.  When the path
from FUSE refers to the top of our filesystem, lowerpath()
returns the "." single-dot path.