convert (hard) link op to use lowerdir fd
As noted in the comments, because we open the lower directory
file descriptor using O_DIRECTORY, we know it's always a directory,
so linkat() should fail as we want it to when asked to create a
hard link on an empty path (i.e., on the top of our filesystem).

In other words, there's no need for the AT_EMPTY_PATH flag as
we definitely don't want to try to create a hard link in this case.