convert rename op to use lowerdir fd
Also note as a to-do comment that if we ever want to support
compilation on BSD, we may need to give up on the direct syscall
and use renameat().  We could throw an error if any flags are
supplied, or (if compiling on Linux only) fall back to using
renameat2() via syscall.