convert xattr ops to use lowerdir fd
Use the fgetxattr() family of functions for our extended
attribute operations.  Unfortunately, these don't take a path;
there's no fgetxattrat() family.

Therefore we use openat() to get a file descriptor on the file itself,
before calling the fgetxattr()-family function.  However, this will hang
indefinitely in an uninterruptable kernel state if the file is
actually a FIFO pipe.  Therefore, we try to check first to prevent
this case, and return ENOTSUP when the file is a pipe.

But, this in turn introduces a small race condition where, in theory,
the file in the underlying lower storage directory could be changed
from regular file to pipe between our check and our trying the openat().

In this case, any client which writes to the pipe will cause the openat()
to return, but it's definitely not ideal and perhaps we can revisit
this in the future.