use autoconf macros for maximal compiler warnings
Rather than roll our own --enable-compile-warnings option and
try to figure out what the best compiler and linker flags for
warnings and errors are, use the convenient m4 macros provided
by the FSF, particularly AX_COMPILER_FLAGS().  See:

https://www.gnu.org/software/autoconf-archive/
https://www.gnu.org/software/autoconf-archive/ax_compiler_flags.html

The default should be to enable the maximum set of warning flags,
so we don't need to change our Docker scripts to get the benefit
of these in CI.

Because of these new compiler warnings, there are a number of
trivial reorganizations of variable declarations, plus a couple
of other fixes for missing enum values in switches and suggested
compiler attributes, etc.