add --with-libfusepkg config option for custom lib
When configuring a user can provide a path to the fuse3.pc
pkg-config file in their custom libfuse build, and the configure
script will extract and set the appropriate compiler and linker
flags using pkg-config.

We prepend the flags to CPPFLAGS and LDFLAGS (unless they happen
to already have been provided, in which case we avoid duplicates).

This should ensure that the subsequent checks for fuse3 headers
and symbols succeeds, checking the custom libfuse path first
before searching for any other libfuse installations.