enhance symlink stat tests
The current symlink stat tests succeed because the primary group of
the user is the first entry in the list returned by the id command,
so while the initial chgrp alters the file referenced by the symlink,
the following stat checks the group of the symlink itself, which is
the first group from the list (regardless of the preceding chgrp)
because the symlink was created with that group by default.

This change enhances the tests to clarify what it tested (default
group of the symlink vs. change to the symlink's reference) and then
check again after running chgrp on the symlink (which will affect
the referenced file, not the symlink itself).

(We could also add tests using chgrp -h, but not all underlying
filesystems may support distinct ownership of symlinks.)