simplify basic filesystem sanity checks for VFS
The VFS API really only pertains to starting and stopping the
mounts, so we don't need a full and exact replication of our t0?? suite
of mirroring tests.

Even the revised t500-vfs-mirror-basic.t in this commit is a bit of
overkill, but at least it's all one file, and does smoke test most
of what we care about, and can be pruned further if desired in the future.