let test library determine tmp paths for cleanup
Instead of hard-coding the test mount path (and optional log output path)
in the Makefile and the test library, create a simple utility script
which allows Make to ask the test library to clean up after itself.

We want "make clean" to remain idempotent, so we need to bypass
both early exit (in the case of a lack of compiled programs) and
creation of test output files until we can run our cleanup and exit.