retry flock at 100ms intervals
There's no good way to time out a flock(2) except with SIGALRM, which in
turn would require us to mess around with pthread sigmasks to ensure
delivery to the correct thread.  We would also require a lock manager to
coordinate multiple threads flock(2)ing and wanting separate
(non-interacting) timeouts on each.

So for now, just use LOCK_NB and waits.