use --retval-file option in handler tests
When the file specified with --retval-file exists, the test
handlers log messages on each event; when the file does not exist,
no messages are output from the test handlers.

This commit has no functional change on the test suite, but once
we add test handlers for projection events, we will need to toggle
test handler output on and off, or else initial projection events
triggered by external system processes may interfere with our tests.

(If additional control proves necessary in the future, we can also
add a list of valid process names or process groups to the control file
and then only report handler events corresponding to those processes.)