remove generic file IO write loop function
We can defer to our callers the implementation details of how they
want to write to the open file descriptor we provide as part of a
file projection event.

In particular, the VFSForGit provider can perform this within its
WriteFileContents() method in the ProjFS.Linux classes which
interact with this library; see github/VFSForGit#6 for the corresponding
changes.