add xattr utilities and use to handle empty flag
Refactor our handling of the "empty" projection flag extended
attribute into two common base utility functions which get or set/remove
xattrs, and a specialized set of wrappers of these common functions
which get, set, or remove our projection flag attribute.

The common base functions catch the the ENOATTR "attribute not found"
error when reading or removing an attribute, and in this case, return
success but reset the caller's 'size' argument to -1 to signal that the
attribute was not found.

We will subsequently use this common pair of base functions to manage
additional, user-provided projection attributes, as will be required
for VFSForGit support.