write user-defined xattrs on file/dir projection
We introduce a common iteration function for parsing and
applying user-provider lists of attributes, and use it to store
such attributes when they are provided during a call to
create a projected file or directory.

The iteration function will also be useful when we add the ability
for callers to subsequently read or reset these attributes after
the initial time of creation.

We store the name/value user attributes as xattrs under our
library's "user.projection.*" namespace, unless they would conflict
with our reserved "user.projection.empty" attribute; we allow
clients to potentially read that attribute, but not overwrite
or remove it.