prevent xattr file ops from overwriting our xattrs
When a normal setxattr() or removexattr() file operation is
invoked via our FUSE event loop, we check if it might alter
or remove one of the attributes we are managing within our
"user.projection.*" namespace, and if so, we deny the request.

While it is still possible for these xattrs to be modified
through changes made directly against our lower storage filesystem,
these checks should prevent inadvertent changes made through
our mount point.