revise event bitmasks to match fanotify/inotify
The *_SELF bitmasks are intended (in inotify use) for events which
occur on the watched inode itself, as opposed to its children.

In our case, virtually all events apply only to descendents of
our mount point, and the mount point itself cannot be easily
deleted/moved/etc. without unmounting, so we should simply use the
appropriate bitmasks for CREATE, MOVE, etc.

We also want to migrate our use of DELETE_SELF to a more appropriate
DELETE_PERM name.