CLOSE_WRITE event notification using fd table
In order to report the CLOSE_WRITE event reliably, and only once
per file descriptor, we deliver it from the FUSE release() operation.
We don't use the flush() operation, as it may be called multiple times
per open file descriptor, or not at all, while release() is always called
exactly once per descriptor.

However, the release() operation's FUSE context no longer has the pid of
the process which owned the file descriptor.  Therefore we track the pid of
the process which opened a file for writing, or most recently closed
that file (thus sending a FUSE flush() operation), and use that pid as our
best effort in our CLOSE_WRITE event notification.

Note that directories can not be opened with write flags, so our
pid tracking and CLOSE_WRITE event notification will not apply to them,
even if they are opened with open(2).