convert kernel thread pid to userspace pid
The pid_t value delivered by FUSE corresponds to the kernel's
lightweight process ID, i.e., the thread ID which would be
returned by a gittid(2) call.  We convert this into the
corresponding process ID, as would be returned by getpid(2),
in part because at the time we deliver certain event notifications,
the thread may have exited.

At present, the only utility for determining the userspace
process ID of a lightweight process (thread) is to parse the
thread's status file in procfs.