parse mount options in test suite
We will need to be able to accept the full range of
(non-test) mount options in the test suite in order to
convert the library to use the optional log file for
error output instead of logging to stderr as it does now.

Since we're adding this capability to the test suite, we
can reuse the test_simple helper for our --initial mount option
functionality tests.