ensure user write perm available before projection
When changing the projection state of files, we may need
to write their contents, and always need to change the
value of their projection state xattr.  Therefore, if we
are unable to reopen the file descriptor with O_WRONLY,
we try to (temporarily) set the user write permission mode
bit, and then retry reopening the file.

Similarly, when directories change projection state, we
need to use fsetxattr() to update their state xattr, and
user write permission on the directory is required to do so.
Hence we perform an fstat() on the directory, and try to
temporarily change the user write permission mode bit, if it
is not set.