enforce user read permission on all inodes
Because we frequently check files and directories to
determine their projection state, we need to always be
able to open them with O_RDONLY in project_dir() and
project_file(), and read their xattrs with fsetxattr().

We therefore force the S_IRUSR mode bit to be set on all
files created by the user within the mount point.  Since
this is a FUSE filesystem without the allow_other mount
option, we can reasonably expect the user to be the owner
of all inodes within the mount.